# gen/docs/swagger/swagger.yaml
definitions:
...
  handlers.ErrorResponse:
    properties:
      error:
        type: string
      trace_id:
        type: string
    type: object
...
  handlers.PasswordChangeRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
      user_id:
        type: string
    required:
    - new_password
    type: object
  handlers.PasswordChangeResponse:
    properties:
      changed_at:
        type: string
      message:
        type: string
      revoked_sessions:
        type: integer
      revoked_tokens:
        type: integer
    type: object
  handlers.PasswordResetConfirmRequest:
    properties:
      code:
        type: string
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    type: object
  handlers.PasswordResetConfirmResponse:
    properties:
      changed_at:
        type: string
      message:
        type: string
      revoked_sessions:
        type: integer
      revoked_tokens:
        type: integer
      user_id:
        type: string
    type: object
  handlers.PasswordResetRequest:
    properties:
      email_or_phone:
        type: string
    required:
    - email_or_phone
    type: object
  handlers.PasswordResetResponse:
    properties:
      delivery:
        type: string
      dev_code:
        type: string
      dev_token:
        description: 'SECURITY: DevToken and DevCode are ONLY exposed in development
          mode'
        type: string
      expires_at:
        type: string
      masked_destination:
        type: string
      message:
        type: string
      request_id:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/password/change:
    post:
      consumes:
      - application/json
      description: Updates the password for the current user or a delegated user when
        permitted.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Password change request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PasswordChangeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Change the password for an authenticated user
      tags:
      - Password
  /api/v1/password/reset/confirm:
    post:
      consumes:
      - application/json
      description: Finalizes the password reset using a token or verification code.
      parameters:
      - description: Password reset confirm request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PasswordResetConfirmResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Complete a password reset
      tags:
      - Password
  /api/v1/password/reset/request:
    post:
      consumes:
      - application/json
      description: Starts the password reset flow and always returns an accepted response
        to avoid account enumeration.
      parameters:
      - description: Password reset request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.PasswordResetResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Initiate a password reset
      tags:
      - Password
...
swagger: "2.0"